maximum roomcount fixed

+++ reservation-service/src/main/java/com/luxury/reservation_service/dto/RoomCountDTO.java
@@ -9,6 +9,6 @@ import lombok.*;
 @Builder
 public class RoomCountDTO {
     private String roomTypeName;
-    private Long availableCount;
+    private Integer availableCount;
 
 }
